Refactor code for clarity and optimization
Generalize fetchTiles and fetchArrows to myGet and add myPost, which
call fetch with 'GET' and 'POST', respectively.
Separate the "tiles" state into two new states, "tilesContent" and
"tilesCoords", in order to change only the coordinates when a tile is
moved and avoid rewriting all the other tile properties.
Make "addArrow" function set "arrowMode" state to false.